apps/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
flask-wtf==1.2.1
flask-limiter==3.7.0
bcrypt==4.1.3
gunicorn==22.0.0
cloudinary==1.40.0
markupsafe==2.1.5
